+++ hyperparameter_hunter/experiments.py
@@ -22,6 +22,8 @@ on whats going on in :mod:`experiment_core`, and its related modules"""
 ##################################################
 # Import Own Assets
 ##################################################
+# noinspection PyProtectedMember
+from hyperparameter_hunter import __version__
 from hyperparameter_hunter.algorithm_handlers import (
     identify_algorithm,
     identify_algorithm_hyperparameters,
@@ -37,6 +39,7 @@ from hyperparameter_hunter.metrics import ScoringMixIn, get_formatted_target_met
 from hyperparameter_hunter.models import model_selector
 from hyperparameter_hunter.recorders import RecorderList
 from hyperparameter_hunter.settings import G
+from hyperparameter_hunter.utils.general_utils import Deprecated
 
 ##################################################
 # Import Miscellaneous Assets
@@ -216,7 +219,8 @@ class BaseExperiment(ScoringMixIn):
             if self.do_raise_repeated is True:
                 self._clean_up()
                 raise RepeatedExperimentError(_ex)
-            G.warn(_ex)
+            G.debug(_ex)
+            G.warn("WARNING: Duplicate experiment!")
 
         self._initialize_random_seeds()
         self._initial_preprocessing()
@@ -224,7 +228,7 @@ class BaseExperiment(ScoringMixIn):
 
         recorders = RecorderList(file_blacklist=G.Env.file_blacklist)
         recorders.format_result()
-        G.log(f'Saving results for Experiment: "{self.experiment_id}"')
+        G.log(f"Saving results for Experiment: '{self.experiment_id}'")
         recorders.save_result()
         self._clean_up()
 
@@ -293,7 +297,7 @@ class BaseExperiment(ScoringMixIn):
         if isinstance(self.test_dataset, pd.DataFrame):
             self.test_input_data = self.test_dataset.copy().loc[:, self.feature_selector]
 
-        G.log("Initial preprocessing stage complete")
+        G.log("Initial preprocessing stage complete", 4)
 
     ##################################################
     # Supporting Methods:
@@ -316,7 +320,7 @@ class BaseExperiment(ScoringMixIn):
             raise EnvironmentInactiveError("")
         if G.Env.current_task is None:
             G.Env.current_task = self
-            G.log(f"Validated Environment with key: '{self.cross_experiment_key}'")
+            G.log(f"Validated Environment:  '{self.cross_experiment_key}'")
         else:
             raise EnvironmentInvalidError("Current experiment must finish before starting another")
 
@@ -331,8 +335,7 @@ class BaseExperiment(ScoringMixIn):
     def _generate_experiment_id(self):
         """Set :attr:`experiment_id` to a UUID"""
         self.experiment_id = str(uuid())
-        G.log("")
-        G.log("Initialized new Experiment with ID: {}".format(self.experiment_id))
+        G.log("Initialized Experiment: '{}'".format(self.experiment_id))
 
     def _generate_hyperparameter_key(self):
         """Set :attr:`hyperparameter_key` to a key to describe the experiment's hyperparameters"""
@@ -347,7 +350,7 @@ class BaseExperiment(ScoringMixIn):
         )
 
         self.hyperparameter_key = HyperparameterKeyMaker(parameters, self.cross_experiment_key)
-        G.log("Generated hyperparameter key: {}".format(self.hyperparameter_key))
+        G.log("Hyperparameter Key:     '{}'".format(self.hyperparameter_key))
 
     def _create_script_backup(self):
         """Create and save a copy of the script that initialized the Experiment if allowed to, and
@@ -363,9 +366,9 @@ class BaseExperiment(ScoringMixIn):
                 except FileNotFoundError:
                     os.makedirs(self.result_paths["script_backup"], exist_ok=False)
                     self._source_copy_helper()
-                G.log("Created backup of file: '{}'".format(self.source_script))
+                G.log("Created source backup:  '{}'".format(self.source_script), 4)
             else:
-                G.log("Skipped creating backup of file: '{}'".format(self.source_script))
+                G.log("Skipped source backup:  '{}'".format(self.source_script), 4)
         #################### Exception Handling ####################
         except AttributeError as _ex:
             if G.Env is None:
@@ -636,6 +639,12 @@ class CrossValidationExperiment(BaseCVExperiment, metaclass=ExperimentMeta):
 ##################################################
 # Other Experiment Classes:
 ##################################################
+@Deprecated(
+    v_deprecate="2.0.0",
+    v_remove="2.1.0",
+    v_current=__version__,
+    details="Superseded by :class:`CrossValidationExperiment`, which is always preferred",
+)
 class RepeatedCVExperiment(BaseCVExperiment, metaclass=ExperimentMeta):
     def __init__(
         self,
@@ -675,6 +684,12 @@ class RepeatedCVExperiment(BaseCVExperiment, metaclass=ExperimentMeta):
             self.folds = RepeatedKFold(**self.cross_validation_params)
 
 
+@Deprecated(
+    v_deprecate="2.0.0",
+    v_remove="2.1.0",
+    v_current=__version__,
+    details="Superseded by :class:`CrossValidationExperiment`, which is always preferred",
+)
 class StandardCVExperiment(BaseCVExperiment, metaclass=ExperimentMeta):
     def __init__(
         self,